Change the top-level Storage; remove the Stat() method in favor for two distinct/specific methods: ContentType and Checksum, for different purposes. Remove unused ChecksumContainer and ObjectInfo. Move the events-related things to a dedicated package.
