+++ pkg/storage/event/event.go
@@ -1,4 +1,4 @@
-package storage
+package event
 
 import (
 	"fmt"
