+++ pkg/storage/event/interfaces.go
@@ -1,0 +1,31 @@
+package event
+
+import (
+	"context"
+	"io"
+
+	"github.com/weaveworks/libgitops/pkg/storage"
+)
+
+// StorageCommon contains the methods that EventStorage adds to the
+// to the normal Storage.
+type StorageCommon interface {
+	// WatchForObjectEvents starts feeding ObjectEvents into the given "into"
+	// channel. The caller is responsible for setting a channel buffering
+	// limit large enough to not block normal operation. An error might
+	// be returned if a maximum amount of watches has been opened already,
+	// e.g. ErrTooManyWatches.
+	WatchForObjectEvents(ctx context.Context, into ObjectEventStream) error
+
+	// Close closes the EventStorage and underlying resources gracefully.
+	io.Closer
+}
+
+// EventStorage is the abstract combination of a normal Storage, and
+// a possiblility to listen for changes to objects as they change.
+// TODO: Maybe we could use some of controller-runtime's built-in functionality
+// for watching for changes?
+type EventStorage interface {
+	storage.Storage
+	StorageCommon
+}
